Driving Course
I built a street for the driving course:

The parked cars, lamps, buildings, and sidewalks are all triggers with
their respective tags - when the player's car collides with them,
they'd register that and change the score - i made a script with the
basics of that, it just needs a way to communicate that info to the
player at the end, when they enter the green "finish area"

The icy patches can now be put over the road - also, feel free to add
more cars or change the cars' positions to where you want to make the
ice challenge.

still no luck with the accelerator/brake pedals.  at this point, we
might have to forget about them so that we can finish.